PlatformInitializer/requirements.txt
aniso8601==9.0.1
asttokens==2.2.1
attrs==22.2.0
azure-common==1.1.28
azure-core==1.26.3
azure-identity==1.12.0
azure-mgmt-compute==29.1.0
azure-mgmt-core==1.3.2
azure-storage-blob==12.15.0
backcall==0.2.0
bcrypt==4.0.1
bidict==0.22.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cryptography==39.0.2
cycler==0.11.0
decorator==5.1.1
docopt==0.6.2
executing==1.2.0
Flask==2.2.3
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
Flask-SocketIO==5.3.3
Flask-SQLAlchemy==3.0.3
fonttools==4.39.3
greenlet==2.0.2
idna==3.4
ipython==8.12.0
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
JsonForm==0.0.2
jsonschema==4.17.3
JsonSir==0.0.2
kafka-python==2.0.2
kazoo==2.9.0
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
msal==1.21.0
msal-extensions==1.0.0
msrest==0.7.1
network==0.1
numpy==1.24.2
oauthlib==3.2.2
packaging==23.1
pandas==1.5.3
paramiko==3.1.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pipreqs==0.4.11
portalocker==2.7.0
prompt-toolkit==3.0.38
psutil==5.9.4
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.15.1
PyJWT==2.6.0
PyNaCl==1.5.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-crontab==2.7.1
python-dateutil==2.8.2
Python-EasyConfig==0.1.7
python-engineio==4.4.0
python-socketio==5.8.0
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
Resource==0.2.1
six==1.16.0
SQLAlchemy==2.0.9
stack-data==0.6.2
storage==0.0.4.3
traitlets==5.9.0
typing_extensions==4.5.0
urllib3==1.26.15
wcwidth==0.2.6
Werkzeug==2.2.3
yarg==0.1.9
